ai-service/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==2.6.3
python-multipart==0.0.9
python-dotenv==1.0.1
httpx==0.27.0
openai==1.14.0

# ML / Image Analysis
numpy
Pillow
torch
torchvision
monai
